+++ databruce/static/admin/js/custom_unfold.js
@@ -1,0 +1,6 @@
+$(document).ready(function () {
+    // console.log($('#submit-row').children('.flex-col-reverse'));
+
+    document.querySelector('#submit-row div div').classList.remove('flex-col-reverse').add('flex-row-reverse');
+    // $('#submit-row').children('.flex-col-reverse').removeClass('flex-col-reverse').addClass('flex-row-reverse');
+});